reporting_service/app/requirements.txt
prophet>=1.1.4
kaleido>=0.2.1
pdfkit>=1.0.0
plotly>=5.0.0
statsmodels>=0.13.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
pandas>=1.3.0
scikit-learn>=1.0.0
groq>=0.3.0
langgraph>=0.0.12
python-multipart
pydantic-settings
scipy>=1.8.0
numpy>=1.20.0
aiofiles>=0.8.0
langchain-community>=0.0.10